refactor: move hooks to hooks.ts file
